fix: Reorganized the codebase for charonctl.
